# Ugly patch for characteristic two in any_root