--- a/src/sage/rings/finite_rings/conway_polynomials.py
+++ b/src/sage/rings/finite_rings/conway_polynomials.py
@@ -136,7 +136,7 @@ class PseudoConwayLattice(WithEqualityById, SageObject):
         sage: # needs sage.rings.finite_rings
         sage: from sage.rings.finite_rings.conway_polynomials import PseudoConwayLattice
         sage: PCL = PseudoConwayLattice(2, use_database=False)
-        sage: PCL.polynomial(3)
+        sage: PCL.polynomial(3)   # random
         x^3 + x + 1
 
     TESTS::
@@ -164,16 +164,16 @@ class PseudoConwayLattice(WithEqualityById, SageObject):
             sage: # needs sage.rings.finite_rings
             sage: from sage.rings.finite_rings.conway_polynomials import PseudoConwayLattice
             sage: PCL = PseudoConwayLattice(3)
-            sage: PCL.polynomial(3)
+            sage: PCL.polynomial(3)  # random
             x^3 + 2*x + 1
 
             sage: # needs sage.rings.finite_rings
             sage: PCL = PseudoConwayLattice(5, use_database=False)
-            sage: PCL.polynomial(12)
+            sage: PCL.polynomial(12)  # random
             x^12 + 4*x^11 + 2*x^10 + 4*x^9 + 2*x^8 + 2*x^7 + 4*x^6 + x^5 + 2*x^4 + 2*x^2 + x + 2
-            sage: PCL.polynomial(6)
+            sage: PCL.polynomial(6)   # random
             x^6 + x^5 + 4*x^4 + 3*x^3 + 3*x^2 + 2*x + 2
-            sage: PCL.polynomial(11)
+            sage: PCL.polynomial(11)  # random
             x^11 + x^6 + 3*x^3 + 4*x + 3
         """
         self.p = p
@@ -215,11 +215,11 @@ class PseudoConwayLattice(WithEqualityById, SageObject):
             sage: # needs sage.rings.finite_rings
             sage: from sage.rings.finite_rings.conway_polynomials import PseudoConwayLattice
             sage: PCL = PseudoConwayLattice(2, use_database=False)
-            sage: PCL.polynomial(3)
+            sage: PCL.polynomial(3)   # random
             x^3 + x + 1
-            sage: PCL.polynomial(4)
+            sage: PCL.polynomial(4)   # random
             x^4 + x^3 + 1
-            sage: PCL.polynomial(60)
+            sage: PCL.polynomial(60)  # random
             x^60 + x^59 + x^58 + x^55 + x^54 + x^53 + x^52 + x^51 + x^48 + x^46 + x^45 + x^42 + x^41 + x^39 + x^38 + x^37 + x^35 + x^32 + x^31 + x^30 + x^28 + x^24 + x^22 + x^21 + x^18 + x^17 + x^16 + x^15 + x^14 + x^10 + x^8 + x^7 + x^5 + x^3 + x^2 + x + 1
         """
         if n in self.nodes:
@@ -239,7 +239,7 @@ class PseudoConwayLattice(WithEqualityById, SageObject):
         # TODO: something like the following
         # gcds = [n.gcd(d) for d in self.nodes.keys()]
         # xi = { m: (...) for m in gcds }
-        xi = {q: self.polynomial(n//q).any_root(K, -n//q, assume_squarefree=True)
+        xi = {q: self.polynomial(n//q).any_root(K, n//q, assume_squarefree=True, assume_distinct_deg=True)
               for q in n.prime_divisors()}
 
         # The following is needed to ensure that in the concrete instantiation
@@ -402,9 +402,9 @@ def _frobenius_shift(K, generators, check_only=False):
         sage: f20 = x^20 + x^19 + x^15 + x^13 + x^12 + x^11 + x^9 + x^8 + x^7 + x^4 + x^2 + x + 1
         sage: f12 = x^12 + x^10 + x^9 + x^8 + x^4 + x^2 + 1
         sage: K.<a> = GF(2^60, modulus='first_lexicographic')
-        sage: x30 = f30.any_root(K)
-        sage: x20 = f20.any_root(K)
-        sage: x12 = f12.any_root(K)
+        sage: x30 = f30.roots(K, multiplicities=False)[0]
+        sage: x20 = f20.roots(K, multiplicities=False)[0]
+        sage: x12 = f12.roots(K, multiplicities=False)[0]
         sage: generators = {2: x30, 3: x20, 5: x12}
         sage: from sage.rings.finite_rings.conway_polynomials import _frobenius_shift, _find_pow_of_frobenius
         sage: _frobenius_shift(K, generators)