… src/sage/rings/polynomial/polynomial_quotient_ring.py
… src/sage/rings/polynomial/polynomial_quotient_ring_element.py
… src/sage/schemes/elliptic_curves/cm.py
… src/sage/schemes/elliptic_curves/ell_field.py