The current method for computing a root of a polynomial using `any_root()` has a mistake in the computation of the trace for even characteristic which means finding a root becomes simply "sample a random element and hope it's a factor". This is incredibly slow as seen in issue #37160.

I tried to simply correct the mistake in the trace, but the resulting algorithm was still incredible slow, so either I have made the incorrect fix or something else is going wrong.

My attempt to fix this was to first implement a method to compute a divisor using Cantor–Zassenhaus:

```py
def compute_divisor(f, d):
    """
    Search for a divisor of the polynomial f of degree d
    """
    F = f.base_ring()
    R = f.parent()
    for _ in range(1000):
        a = R.random_element()
        if a.is_zero():
            continue
        a = a.monic()
        b = sum([pow(a, 2**i, f) for i in range(d * F.degree())])
        g = f.gcd(b)
        if not (g.is_one() or f == g):
            return g
    raise ValueError("no divisor found")
```

Testing this, I think it works as I would expect and I find divisors very quickly:

```py
sage: R.<x> = GF(2^8)[]
sage: f = R.random_element(degree=6); len(f.roots())
3
sage: compute_divisor(f, 1)
x + z8^4 + z8^3 + 1
sage: compute_divisor(f, 2)
x^3 + (z8^7 + z8^5 + z8^2 + 1)*x^2 + (z8^7 + z8^5 + z8^3 + z8^2)*x + z8^7 + z8^6 + z8^2 + z8 + 1
sage: compute_divisor(f, 3)
x + z8^7 + z8^6 + z8^5 + z8^4 + z8^3 + 1
sage: f.factor()
(z8^7 + z8^2 + z8 + 1) * (x + z8^4 + z8^3 + 1) * (x + z8^6 + z8^2 + 1) * (x + z8^7 + z8^6 + z8^5 + z8^4 + z8^3 + 1) * (x^3 + (z8^6 + z8^3 + z8^2 + 1)*x^2 + (z8^6 + z8^4 + z8^2 + z8)*x + z8^7 + z8^4 + z8^2 + z8)
```

So the natural thing I thought was to patch the `any_roots()` to replace:

```python
T = T.monic()
C = T
for i in range(degree-1):
    C = T + pow(C,q,self)
```

To instead iterate over $\sum_{0}^{dk - 1} T^{2^i} \mod f$ with 

```python
C = sum([pow(T, 2**i, self) for i in range(d * self.base_ring().degree())])
```

Which can be computed with less redundancy as:

```python
C = T
TT = T
for _ in range(d * self.base_ring().degree() - 1):
    TT = pow(TT, 2, self)
    C += TT
```

But when I rebuilt this, I noticed that it was still taking 10s of seconds to find a root.

I then noticed an edge case for:

```python
k = self.degree() // degree
if k == 1:
    try:
        return self.roots(ring, multiplicities=False)[0] # is there something better to do here?
    except IndexError:
        raise ValueError("no roots F %s" % self)
```

so for now, I include characteristic 2 into this edge case and comment out the C-Z technique which is either broken or just incredibly slow?

```py
if k == 1 or q % 2 == 0:
    # Is there something better to do here for k = 1?
    roots = self.roots(ring, multiplicities=False)
    if roots:
        from sage.misc.prandom import choice
        return choice(roots)
    else:
        raise ValueError(f"no roots F {self}" % self)
```

This is MUCH faster, but it's disappointing that the maths seems to be broken, so I would rather find a way to fix this. Maybe there's a silly bug elsewhere I have missed?

### Before PR

```python
┌────────────────────────────────────────────────────────────────────┐
│ SageMath version 10.2, Release Date: 2023-12-03                    │
│ Using Python 3.11.4. Type "help()" for help.                       │
└────────────────────────────────────────────────────────────────────┘
sage: set_random_seed(0)
....: K.<z8> = GF(2^8)
....: R.<x> = K[]
....: u = R.irreducible_element(2)
....: K_ext = K.extension(modulus=u, names="a")
....: R_ext.<y_ext> = K_ext[]
....: poly = R_ext.random_element(2).monic()
....: %time poly.roots()
CPU times: user 20.5 s, sys: 19.9 ms, total: 20.5 s
Wall time: 20.5 s
[]
```

### After PR

```py
┌────────────────────────────────────────────────────────────────────┐
│ SageMath version 10.3.beta6, Release Date: 2024-01-21              │
│ Using Python 3.11.4. Type "help()" for help.                       │
└────────────────────────────────────────────────────────────────────┘
┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
┃ Warning: this is a prerelease version, and it may be unstable.     ┃
┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
sage: set_random_seed(0)
....: K.<z8> = GF(2^8)
....: R.<x> = K[]
....: u = R.irreducible_element(2)
....: K_ext = K.extension(modulus=u, names="a")
....: R_ext.<y_ext> = K_ext[]
....: poly = R_ext.random_element(2).monic()
....: %time poly.roots()
CPU times: user 85.8 ms, sys: 7.37 ms, total: 93.2 ms
Wall time: 117 ms
[]
```