@GiacomoPope 2024-01-26
Also, `any_roots()` seems very complicated to me, but maybe it's necessary. I suppose this could be an opportunity to refactor the whole function.