@GiacomoPope 2024-01-26
I think I will try and refactor `any_root` as the more I look at it, the less I understand it and think it can be simplified.