@GiacomoPope 2024-01-29
With some small tweaks to the doctests (root finding is not deterministic anymore, which i guess it never was due to using CZ, but the doctests assumed it was for certain cases?)

The `any_root()` now seems to behave as it used to. Although making this happen made the code uglier, so I think it would be good to deprecate certain API. I think the case with `degree` and `ring` both not None is the weirdest of the choices and could be avoided, but i need opinions.