@GiacomoPope 2024-01-31
New push should get this "finished" or totally ready for review.

## Summary

- Rewrote `any_root()` as the previous implementation was too complex and had many bugs
- Now isomorphisms for `GF(2^k)` actually work, rather than randomly sampling roots until a factor is found! For GF(2^8) we now can compute isomorphisms in 100ms instead of 20-40 seconds!!
- Deprecated allowing negative degree to make an assumption about the factorisation of the input. Instead we have a new `bool` for this.
- Instead of falling back to factorising and taking roots when `degree > 0`, we now recursively call `any_root()` on the degree `degree` irreducible polynomial over an extension ring.
- Added this new bool throughout the codebase and removed all negative degree inputs.
- New implementation is now no longer deterministic. The old one was "sometimes, depending on optional parameters....". This caused an issue with some doctests which I added # random tags to.