@GiacomoPope 2024-02-02
Thanks for all the reviews and advice on refactoring. (I think) I even feel happy with how this function works haha!