@vbraun 2024-02-04
I'm getting
```
sage -t --warn-long 40.4 --random-seed=123 src/sage/rings/polynomial/polynomial_quotient_ring_element.py
**********************************************************************
File "src/sage/rings/polynomial/polynomial_quotient_ring_element.py", line 717, in sage.rings.polynomial.polynomial_quotient_ring_element.PolynomialQuotientRingElement.minpoly
Failed example:
    ext(u + 1).minpoly()  # indirect doctest                              # needs sage.modules
Expected:
    x^3 + (396*i + 428)*x^2 + (80*i + 39)*x + 9*i + 178
Got:
    x^3 + (35*i + 428)*x^2 + (351*i + 39)*x + 422*i + 178
**********************************************************************
1 item had failures:
   1 of  19 in sage.rings.polynomial.polynomial_quotient_ring_element.PolynomialQuotientRingElement.minpoly
    [160 tests, 1 failure, 0.51 s]
----------------------------------------------------------------------
sage -t --warn-long 40.4 --random-seed=123 src/sage/rings/polynomial/polynomial_quotient_ring_element.py  # 1 doctest failed
----------------------------------------------------------------------
```