@GiacomoPope 2024-02-04
Thank you for reporting the error -- I'll see whether I can find the root of this, I assume it'll be again that `any_root()` is no longer deterministic