@vbraun 2024-02-05
I can live with it being nondeterministic 

Ideally we'd not just slap `# random` on tests but also check something that is true (like the `g(f(t+1)) == t+1`), both as clarification for the docs and to guard against regressions.