@GiacomoPope 2024-02-05
Right, I could look at including checks like these in more places, at least for the `hom_*` files where the `# random` has appeared the most. If it helps this PR I am happy to take some time to rewrite some of the `hom` doctests to be better "tests" in light of the change to non-determinism.

Would changes like this be needed in other files too, or simply `hom_finite_field*`? For the `PCL.polynomial(3)` I am afraid I have no experience here and do not know "good" doctests