@vbraun 2024-02-05
If its easy enough to add then go for it, but if not then not.