@GiacomoPope 2024-02-05
Modified the tests to check things like `f(a + b) == f(a) + f(b)` and also made the section tests more explicit by picking random `a` and then ensuring `g(f(a)) == a`.

Did not touch the `PCL.polynomial` code again, as I'm just not sure what to do.